Update UI + Add more controls.

+++ src/main/java/lu/itrust/business/TS/controller/ControllerProfile.java
@@ -152,7 +152,7 @@ public class ControllerProfile {
 	}
 
 	@RequestMapping(value = "/Sqlite/{id}/Delete", method = RequestMethod.POST, headers = ACCEPT_APPLICATION_JSON_CHARSET_UTF_8)
-	public @ResponseBody String deleteSqlite(@PathVariable Integer id, Principal principal, Locale locale) throws Exception {
+	public @ResponseBody String deleteSqlite(@PathVariable Long id, Principal principal, Locale locale) throws Exception {
 		UserSQLite userSQLite = serviceUserSqLite.getByIdAndUser(id, principal.getName());
 		if (userSQLite == null)
 			return JsonMessage.Error(messageSource.getMessage("error.resource.not.found", null, "Resource cannot be found", locale));
@@ -253,7 +253,7 @@ public class ControllerProfile {
 	 * @throws Exception
 	 */
 	@RequestMapping("/Sqlite/{id}/Download")
-	public String downloadSqlite(@PathVariable Integer id, Principal principal, HttpServletResponse response, Locale locale) throws Exception {
+	public String downloadSqlite(@PathVariable Long id, Principal principal, HttpServletResponse response, Locale locale) throws Exception {
 
 		// get user file by given file id and username
 		UserSQLite userSqLite = serviceUserSqLite.getByIdAndUser(id, principal.getName());
